adding export fiddle script
